Move react-select into own package
As it can be replaced by
https://github.com/vivid-planet/react-admin/pull/256 and
https://github.com/vivid-planet/react-admin/pull/250
