+++ packages/react-admin-react-select/.npmignore
@@ -1,0 +1,1 @@
+src
