+++ packages/react-admin-react-select/package.json
@@ -1,0 +1,35 @@
+{
+    "name": "@vivid-planet/react-admin-react-select",
+    "version": "0.1.0",
+    "main": "lib/index.js",
+    "types": "lib/index.d.ts",
+    "files": [
+        "lib"
+    ],
+    "dependencies": {
+        "classnames": "^2.2.6"
+    },
+    "devDependencies": {
+        "@vivid-planet/react-admin": "~0.1.0"
+    },
+    "peerDependencies": {
+        "@material-ui/core": "^4.1.3",
+        "@material-ui/icons": "^4.2.1",
+        "@vivid-planet/react-admin": "~0.1.0",
+        "final-form": "^4.16.1",
+        "react": "^16.8",
+        "react-final-form": "^6.3.1",
+        "react-select": "^3.0.4"
+    },
+    "scripts": {
+        "prebuild": "rimraf lib",
+        "build": "tsc --project ./tsconfig.json",
+        "start": "tsc --project ./tsconfig.json --watch",
+        "prepare": "npm run build",
+        "lint:tsc": "tsc --noEmit"
+    },
+    "publishConfig": {
+        "access": "public"
+    },
+    "license": "BSD-2-Clause"
+}
