+++ packages/react-admin-react-select/src/FinalFormReactSelectStaticOptions.tsx
@@ -3,7 +3,7 @@ import { FieldInputProps, FieldRenderProps } from "react-final-form";
 import { Props as ReactSelectProps } from "react-select/base";
 import { OptionsType } from "react-select/src/types";
 
-import { FinalFormReactSelect } from "./ReactSelect";
+import { FinalFormReactSelect } from "./FinalFormReactSelect";
 
 interface IOptionType {
     value: string;
