+++ packages/react-admin-react-select/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./ReactSelect";
+export * from "./FinalFormReactSelect";
+export * from "./FinalFormReactSelectStaticOptions";
