+++ packages/react-admin/package.json
@@ -7,7 +7,6 @@
         "lib"
     ],
     "dependencies": {
-        "classnames": "^2.2.6",
         "debounce": "^1.2.0",
         "lodash.isequal": "^4.5.0",
         "query-string": "^6.8.1",
@@ -35,7 +34,6 @@
         "react-intl": "^5.10.0",
         "react-router": "^5.1.2",
         "react-router-dom": "^5.1.2",
-        "react-select": "^3.0.4",
         "styled-components": "^4.3.2"
     },
     "scripts": {
