+++ packages/react-admin/src/form/index.ts
@@ -4,8 +4,6 @@ export * from "./FieldContainerLabelAbove";
 export * from "./FormPaper";
 export * from "./FormSection";
 export * from "./Input";
-export * from "./ReactSelect";
-export * from "./ReactSelectStaticOptions";
 export * from "./themeAugmentation";
 export * from "./Checkbox";
 export * from "./Radio";
