+++ packages/react-admin/src/mui/index.ts
@@ -1,6 +1,5 @@
 export * from "./breadcrumbs";
 export * from "./menu";
-export * from "./reactSelect";
 export * from "./ThemeProvider";
 export * from "./MasterLayout";
 export * from "./styled-components";
