--- packages/react-admin/src/mui/reactSelect/index.ts
@@ -1,2 +1,0 @@
-export * from "./Select";
-export * from "./themeAugmentation";
